Add self-registration. Add ban command.

+++ discord/commands/createuser.go
@@ -38,7 +38,11 @@ func createuserCommand(s *discordgo.Session, m *discordgo.Message) {
 		return
 	}
 
-	generated, ok := db.CreateUser(s, m, member)
+	generated, ok := db.CreateUser(member)
+	if generated == "" || !ok {
+		_, _ = s.ChannelMessageSend(m.ChannelID, "Error! Please try again later.")
+		return
+	}
 
 	dm, _ := s.UserChannelCreate(member.User.ID)
 	_, err = s.ChannelMessageSend(dm.ID, "Your Turtl API key is: `"+generated+"`. Please do not lose it or give it to anyone else.\n\nYou can generate a .sxcu (configuration) file by going to <#737767470789820496> and typing `+sxcu` (your API kill will be filled automatically).")
